Что-то пошло не так, и мои предыдущие коммиты накрылись тазом. Скопировал проект. Накидал кнопки в TableLayout, заинклюдил в Linear, добавил окно для вывода результата, поменял дизайн кнопок и их расположение.
