--- Calc/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Thu Jun 03 18:11:40 YEKT 2021
-distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.7.1-bin.zip
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
