+++ Calculator/gradle.properties
